cosas de salperro

+++ E_Commerce_VS/Controllers/ControladorProductoCarrito.cs
@@ -25,21 +25,20 @@ namespace E_Commerce_VS.Controllers
             return await _context.ProductoCarritos.Include(pc => pc.Producto).ToListAsync();
         }
 
+        // Eliminar un producto del carrito
         [HttpPut("eliminarproductocarrito")]
         public async Task<IActionResult> DeleteProduct([FromBody] DeleteProductDto request)
         {
-            // Buscar el carrito asociado al usuario
-            var carrito = await _context.Carritos
-                .FirstOrDefaultAsync(c => c.UserId == request.UserId);
-
-            // Buscar el producto en el carrito
             var productoCarrito = await _context.ProductoCarritos
-                .FirstOrDefaultAsync(pc => pc.ProductoId == request.ProductId && pc.CarritoId == carrito.Id);
+                .FirstOrDefaultAsync(pc => pc.ProductoId == request.ProductId && pc.CarritoId == request.CarritoId);
 
-            // Eliminarlo
+            if (productoCarrito == null)
+            {
+                return NotFound("Producto no encontrado en el carrito.");
+            }
+
             _context.ProductoCarritos.Remove(productoCarrito);
             await _context.SaveChangesAsync();
-
             return Ok("Producto eliminado del carrito.");
         }
 
@@ -62,22 +61,11 @@ namespace E_Commerce_VS.Controllers
         }
 
         // Obtener productos de un carrito específico
-        [HttpGet("productosCarrito/{userId}")]
-        public async Task<IActionResult> GetProductosCarritoByUserId(int userId)
+        [HttpGet("productosCarrito/{carritoId}")]
+        public async Task<IActionResult> GetProductosCarrito(int carritoId)
         {
-            // Verificar si el carrito existe para el usuario
-            var carrito = await _context.Carritos
-                .Include(c => c.ProductoCarrito)
-                .FirstOrDefaultAsync(c => c.UserId == userId);
-
-            if (carrito == null)
-            {
-                return NotFound($"No se encontró un carrito para el usuario con ID {userId}.");
-            }
-
-            // Obtener los productos asociados al carrito
             var productos = await _context.ProductoCarritos
-                .Where(pc => pc.CarritoId == carrito.Id)
+                .Where(pc => pc.CarritoId == carritoId)
                 .Include(pc => pc.Producto)
                 .Select(pc => new
                 {
